PTSD-1483 updated to pa-common SNAPSHOT146615 and hw-mongo SNAPSHOT146659, for both smrt-basecaller and pa-cal. new versions:

+++ src/basecaller/appModules/TraceFileDataSource.cpp
@@ -488,7 +488,7 @@ std::vector<DataSourceBase::UnitCellProperties> TraceFileDataSource::GetUnitCell
     if (mode_ == Mode::Replication)
     {
         assert(replicationDims_.has_value());
-        assert(replicationDims_->first * replicationDims_->second == numZmw);
+        assert(replicationDims_->first * replicationDims_->second == (int)numZmw);
 
         for (uint32_t i = 0; i < features.size(); ++i)
         {
